Move protocol struct to namespace caf
