+++ libcaf_core/src/protocol.cpp
@@ -16,55 +16,16 @@
  * http://www.boost.org/LICENSE_1_0.txt.                                      *
  ******************************************************************************/
 
-#pragma once
-
-#include <cstddef>
-#include <string>
-
-#include "caf/meta/type_name.hpp"
+#include "caf/protocol.hpp"
 
 namespace caf {
-namespace io {
-namespace network {
 
-/// Bundles protocol information for network and transport layer communication.
-struct protocol {
-  /// Denotes a network protocol, i.e., IPv4 or IPv6.
-  enum network {
-    ipv4,
-    ipv6
-  };
-  /// Denotes a transport protocol, i.e., TCP or UDP.
-  enum transport {
-    tcp,
-    udp
-  };
-  transport trans;
-  network net;
-};
-
-/// @relates protocol::transport
-inline std::string to_string(protocol::transport x) {
-  return x == protocol::tcp ? "TCP" : "UDP";
+std::string to_string(const protocol& x) {
+  std::string result;
+  result += to_string(x.trans);
+  result += "/";
+  result += to_string(x.net);
+  return result;
 }
 
-/// @relates protocol::network
-inline std::string to_string(protocol::network x) {
-  return x == protocol::ipv4 ? "IPv4" : "IPv6";
-}
-
-/// @relates protocol
-template <class Inspector>
-typename Inspector::result_type inspect(Inspector& f, protocol& x) {
-  return f(meta::type_name("protocol"), x.trans, x.net);
-}
-
-/// Converts a protocol into a transport/network string representation, e.g.,
-/// "TCP/IPv4".
-/// @relates protocol
-std::string to_string(const protocol& x);
-
-} // namespace network
-} // namespace io
 } // namespace caf
-
