+++ libcaf_io/caf/io/all.hpp
@@ -29,7 +29,6 @@
 #include "caf/io/middleman_actor.hpp"
 #include "caf/io/system_messages.hpp"
 
-#include "caf/io/network/protocol.hpp"
 #include "caf/io/network/interfaces.hpp"
 #include "caf/io/network/multiplexer.hpp"
 #include "caf/io/network/test_multiplexer.hpp"
