+++ libcaf_io/caf/io/network/multiplexer.hpp
@@ -24,6 +24,7 @@
 
 #include "caf/extend.hpp"
 #include "caf/expected.hpp"
+#include "caf/protocol.hpp"
 #include "caf/resumable.hpp"
 #include "caf/make_counted.hpp"
 #include "caf/execution_unit.hpp"
@@ -32,7 +33,6 @@
 #include "caf/io/accept_handle.hpp"
 #include "caf/io/connection_handle.hpp"
 
-#include "caf/io/network/protocol.hpp"
 #include "caf/io/network/ip_endpoint.hpp"
 #include "caf/io/network/native_socket.hpp"
