+++ libcaf_io/src/middleman.cpp
@@ -417,8 +417,7 @@ void middleman::init(actor_system_config& cfg) {
   cfg.group_module_factories.emplace_back(gfactory);
   // logging not available at this stage
   // add I/O-related types to config
-  cfg.add_message_type<network::protocol>("@protocol")
-     .add_message_type<network::address_listing>("@address_listing")
+  cfg.add_message_type<network::address_listing>("@address_listing")
      .add_message_type<network::receive_buffer>("@receive_buffer")
      .add_message_type<new_data_msg>("@new_data_msg")
      .add_message_type<new_connection_msg>("@new_connection_msg")
